+++ gpaw/lrtddft/omega_matrix.py
@@ -92,8 +92,7 @@ class OmegaMatrix:
                 self.xc = XC(xc)
             else:
                 self.xc = xc
-            self.xc.initialize(self.paw.density, self.paw.hamiltonian,
-                               wfs, self.paw.occupations)
+            self.xc.initialize(self.paw.density, self.paw.hamiltonian, wfs)
 
             # check derivativeLevel
             if derivativeLevel is None: