+++ gpaw/test/test_fixocc.py
@@ -15,19 +15,19 @@ def test_fixocc():
 
     H2 = Cluster(molecule('H2'))
     H2.minimal_box(box, h)
-    convergence = {'energy':0.01, 'eigenstates':1.e-3, 'density':1.e-2}
+    convergence = {'energy': 0.01, 'eigenstates': 0.001, 'density': 0.01}
 
     if 1:
         # test ZeroKelvin vs FixedOccupations
         c = GPAW(h=h, nbands=nbands,
-                 occupations=ZeroKelvin(True),
+                 occupations=ZeroWidth(),
                  convergence=convergence,
                  txt=txt)
         H2.calc = c
         E_zk = H2.get_potential_energy()
 
         c = GPAW(h=h, nbands=nbands,
-                 occupations=FixedOccupations([[2, 0]]),
+                 occupations=FixedOccupations([[1, 0]]),
                  convergence=convergence,
                  txt=txt)
         H2.calc = c
@@ -38,7 +38,7 @@ def test_fixocc():
     if 1:
         # test spin-paired vs spin-polarized
         c = GPAW(h=h, nbands=nbands,
-                 occupations=FixedOccupations([[1,1]]),
+                 occupations=FixedOccupations([[1, 1]]),
                  convergence=convergence,
                  txt=txt)
         H2.calc = c