+++ gpaw/test/test_occupations.py
@@ -40,6 +40,6 @@ def test_occupations(func):
 @pytest.mark.ci
 def test_occupation_obj(e_kn, w_k, ne):
     for occ in [FermiDirac(0.1), ZeroWidth()]:
-        f_kn, fl, s = occ(ne, e_kn, w_k)
+        f_kn, fl, s = occ.calculate(ne, e_kn, w_k)
         print(f_kn, fl, s)
         assert f_kn.sum(1).dot(w_k) == pytest.approx(ne, abs=1e-14)